+++ src/gondarwizard.h
@@ -123,7 +123,7 @@ class GondarWizard : public QWizard {
 
   int nextId() const override;
   void postError(const QString& error);
-  double getRunTime();
+  qint64 getRunTime();
   // There's an elaborate state-sharing solution via the 'field' mechanism
   // supported by QWizard.  I found the logic for that to be easy for sharing
   // some data types and convoluted for others.  In this case, a later page
